worker/requirements.txt
python-dotenv
pyfiglet
requests
flask
flask-classy
bson
selenium
pymongo
slimit
beautifulsoup4
pyyaml
docker
influxdb
hazelcast-python-client
kafka-python
